requirements.txt
fastapi
uvicorn
pydantic
passlib[bcrypt]
bcrypt==4.0.1
python-jose[cryptography]
python-multipart
